+++ hyperdjango/__init__.py
@@ -1,4 +1,4 @@
-from hyperdjango.actions import Actions, action
+from hyperdjango.actions import Actions, Event, action
 from hyperdjango.page import HyperActionMixin, HyperPageTemplate, HyperView, Page
 from hyperdjango.shortcuts import render_template_block, render_template_page
 
@@ -8,6 +8,7 @@ __all__ = [
     "HyperView",
     "Page",
     "Actions",
+    "Event",
     "action",
     "render_template_page",
     "render_template_block",